@@ -26,43 +26,47 @@
   "repository": "Folkloreatelier/react-scene",
   "homepage": "https://github.com/Folkloreatelier/react-scene#readme",
   "bugs": "https://github.com/Folkloreatelier/react-scene/issues",
-  "main": "dist/react-scene.js",
+  "main": "lib/index.js",
   "scripts": {
-    "clean:dist": "rm -rf ./dist",
+    "clean:lib": "rm -rf ./lib",
     "clean:tmp": "rm -rf ./.tmp",
-    "clean": "npm run clean:tmp && npm run clean:dist",
-    "mkdir:dist": "mkdir -p ./dist",
+    "clean": "npm run clean:tmp && npm run clean:lib",
+    "mkdir:lib": "mkdir -p ./lib",
     "mkdir:tmp": "mkdir -p ./.tmp",
-    "mkdir": "npm run mkdir:tmp && npm run mkdir:dist",
+    "mkdir": "npm run mkdir:tmp && npm run mkdir:lib",
     "build:files": "npm run clean && npm run mkdir",
     "build": "npm run test && npm run build:files && npm run build:js && npm run clean:tmp",
     "watch": "npm run clean:tmp && npm run mkdir:tmp",
     "browsersync": "node build/browsersync.js",
     "server": "concurrently \"npm run watch\" \"npm run browsersync\"",
-    "test": "mocha",
-    "jshint:dist": "jshint src/**.js",
-    "jshint": "npm run jshint:dist",
+    "test": "jest",
+    "eslint:dist": "eslint src/**.js",
+    "eslint": "npm run eslint:dist",
     "jscs": "jscs src/**.js",
-    "webpack:dist": "webpack --config build/webpack.config.build.js",
-    "webpack:min": "webpack --config build/webpack.config.min.js",
+    "babel": "BABEL_ENV=production babel -d lib/ src/",
+    "webpack:dist": "webpack --env=prod --config build/webpack.config.js",
+    "webpack:min": "webpack --env=min --config build/webpack.config.js",
     "webpack": "npm run webpack:dist && npm run webpack:min",
-    "scripts:dist": "npm run webpack",
+    "scripts:dist": "npm run webpack && npm run babel",
     "scripts": "npm run scripts:dist",
-    "build:js": "npm run jshint && npm run scripts",
+    "build:js": "npm run eslint && npm run scripts",
     "release": "build/release.sh"
   },
   "dependencies": {
+    "change-case": "^3.0.1",
     "hoist-non-react-statics": "^1.2.0"
   },
   "peerDependencies": {
-    "react": "^0.14.0 || ^15.0.0"
+    "react": "^15.0.0"
   },
   "devDependencies": {
     "autoprefixer": "^6.4.1",
+    "babel-core": "^6.22.1",
+    "babel-eslint": "^7.1.1",
     "babel-loader": "^6.2.5",
     "babel-preset-es2015": "^6.14.0",
     "babel-preset-react": "^6.11.1",
-    "babel-preset-stage-0": "^6.5.0",
+    "babel-preset-stage-0": "^6.22.0",
     "brfs": "^1.4.3",
     "browser-sync": "^2.15.0",
     "bs-fullscreen-message": "^1.1.0",
@@ -73,14 +77,19 @@
     "cssnano": "^3.7.4",
     "customizr": "^1.0.0-alpha",
     "enzyme": "^2.4.1",
+    "eslint": "^3.15.0",
+    "eslint-config-airbnb": "^14.1.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "eslint-plugin-react": "^6.9.0",
     "expose-loader": "^0.7.1",
     "html-loader": "^0.4.3",
     "imagemin-cli": "^3.0.0",
     "imports-loader": "^0.6.5",
+    "jest": "^18.1.0",
     "jsdom": "^9.5.0",
     "jshint": "^2.9.3",
     "json-loader": "^0.5.4",
-    "mocha": "^3.0.2",
     "postcss-cli": "^2.6.0",
     "proxy-middleware": "^0.15.0",
     "raw-loader": "^0.5.1",
@@ -94,7 +103,8 @@
     "style-loader": "^0.13.1",
     "svg-react-loader": "^0.3.7",
     "transform-loader": "^0.2.3",
-    "webpack": "^1.13.2",
-    "webpack-dev-middleware": "^1.6.1"
+    "webpack": "^2.2.1",
+    "webpack-dev-middleware": "^1.10.0",
+    "webpack-merge": "^2.6.1"
   }
 }
