@@ -1,4 +1,3 @@
+import createReactScene from './createReactScene';
 
-var createReactScene = require('./createReactScene');
-
-module.exports = createReactScene();
+export default createReactScene();
