@@ -1,32 +1,22 @@
-var ReactScene = require('./ReactScene');
-var hoistStatics = require('hoist-non-react-statics');
+import React from 'react';
+import hoistStatics from 'hoist-non-react-statics';
 
-function getDisplayName(WrappedComponent)
-{
-    return WrappedComponent.displayName || WrappedComponent.name || 'Component';
-}
+const getDisplayName = WrappedComponent => (
+    WrappedComponent.displayName || WrappedComponent.name || 'Component'
+);
 
-module.exports = function(WrappedComponent)
-{
-    var WithScene = React.createClass({
-        
-        contextTypes: {
-            scene: React.PropTypes.object
-        },
-        
-        render: function()
-        {
-            /* jshint ignore:start */
-            return (
-                <WrappedComponent {...this.props} scene={this.context.scene} />
-            );
-            /* jshint ignore:end */
-        }
-        
-    });
-    
+export default (WrappedComponent) => {
+    const contextTypes = {
+        scene: React.PropTypes.object,
+    };
+
+    const WithScene = (props, context) => (
+        <WrappedComponent {...props} scene={context.scene} />
+    );
+
+    WithScene.contextTypes = contextTypes;
     WithScene.displayName = `withScene(${getDisplayName(WrappedComponent)})`;
     WithScene.WrappedComponent = WrappedComponent;
-    
+
     return hoistStatics(WithScene, WrappedComponent);
 };
